moved version endpoints and tests into separate files
